重新支持pc

+++ libs/core/extension/resource/Resource.js
@@ -1,20 +1,31 @@
-/**
- * Copyright (c) Egret-Labs.org. Permission is hereby granted, free of charge,
- * to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom
- * the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
- * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+//////////////////////////////////////////////////////////////////////////////////////
+//
+//  Copyright (c) 2014-2015, Egret Technology Inc.
+//  All rights reserved.
+//  Redistribution and use in source and binary forms, with or without
+//  modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//     * Neither the name of the Egret nor the
+//       names of its contributors may be used to endorse or promote products
+//       derived from this software without specific prior written permission.
+//
+//  THIS SOFTWARE IS PROVIDED BY EGRET AND CONTRIBUTORS "AS IS" AND ANY EXPRESS
+//  OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+//  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+//  IN NO EVENT SHALL EGRET AND CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+//  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+//  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;LOSS OF USE, DATA,
+//  OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+//  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+//  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+//  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+//////////////////////////////////////////////////////////////////////////////////////
 var RES;
 (function (RES) {
     /**
@@ -135,10 +146,11 @@ var RES;
      * 销毁单个资源文件或一组资源的缓存数据,返回是否删除成功。
      * @method RES.destroyRes
      * @param name {string} 配置文件中加载项的name属性或资源组名
+     * @param force {boolean} 销毁一个资源组时其他资源组有同样资源情况资源是否会被删除，默认值true
      * @returns {boolean}
      */
-    function destroyRes(name) {
-        return instance.destroyRes(name);
+    function destroyRes(name, force) {
+        return instance.destroyRes(name, force);
     }
     RES.destroyRes = destroyRes;
     /**
@@ -570,11 +582,13 @@ var RES;
          * 销毁单个资源文件或一组资源的缓存数据,返回是否删除成功。
          * @method RES.destroyRes
          * @param name {string} 配置文件中加载项的name属性或资源组名
+         * @param force {boolean} 销毁一个资源组时其他资源组有同样资源情况资源是否会被删除，默认值true
          * @returns {boolean}
          */
-        __egretProto__.destroyRes = function (name) {
+        __egretProto__.destroyRes = function (name, force) {
+            if (force === void 0) { force = true; }
             var group = this.resConfig.getRawGroupByName(name);
-            if (group) {
+            if (group && group.length > 0) {
                 var index = this.loadedGroups.indexOf(name);
                 if (index != -1) {
                     this.loadedGroups.splice(index, 1);
@@ -582,10 +596,14 @@ var RES;
                 var length = group.length;
                 for (var i = 0; i < length; i++) {
                     var item = group[i];
-                    item.loaded = false;
-                    var analyzer = this.getAnalyzerByType(item.type);
-                    analyzer.destroyRes(item.name);
-                    this.removeLoadedGroupsByItemName(item.name);
+                    if (!force && this.isResInLoadedGroup(item.name)) {
+                    }
+                    else {
+                        item.loaded = false;
+                        var analyzer = this.getAnalyzerByType(item.type);
+                        analyzer.destroyRes(item.name);
+                        this.removeLoadedGroupsByItemName(item.name);
+                    }
                 }
                 return true;
             }
@@ -617,6 +635,21 @@ var RES;
                     }
                 }
             }
+        };
+        __egretProto__.isResInLoadedGroup = function (name) {
+            var loadedGroups = this.loadedGroups;
+            var loadedGroupLength = loadedGroups.length;
+            for (var i = 0; i < loadedGroupLength; i++) {
+                var group = this.resConfig.getRawGroupByName(loadedGroups[i]);
+                var length = group.length;
+                for (var j = 0; j < length; j++) {
+                    var item = group[j];
+                    if (item.name == name) {
+                        return true;
+                    }
+                }
+            }
+            return false;
         };
         /**
          * 设置最大并发加载线程数量，默认值是2.
